backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scipy
python-dotenv
pydantic
scikit-learn
